requirements.txt
streamlit
openai
pandas
python-dotenv
SpeechRecognition
audio-recorder-streamlit
